refactor single module into package
