+++ pdm/installers/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+from pdm.installers.installers import Installer
+from pdm.installers.synchronizers import Synchronizer
